requirements/requirements.dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/requirements.dev.in -o requirements/requirements.dev.txt
amqp==5.3.1
    # via
    #   -r requirements/requirements.txt
    #   kombu
annotated-types==0.7.0
    # via pydantic
asgiref==3.8.1
    # via
    #   -r requirements/requirements.txt
    #   django
    #   django-cors-headers
    #   django-stubs
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via
    #   -r requirements/requirements.txt
    #   jsonschema
    #   referencing
authlib==1.5.1
    # via safety
billiard==4.2.1
    # via
    #   -r requirements/requirements.txt
    #   celery
celery==5.4.0
    # via -r requirements/requirements.txt
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   -r requirements/requirements.txt
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   nltk
    #   safety
    #   typer
click-didyoumean==0.3.1
    # via
    #   -r requirements/requirements.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements/requirements.txt
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements/requirements.txt
    #   celery
coverage==7.6.12
    # via
    #   -r requirements/requirements.txt
    #   pytest-cov
cryptography==44.0.2
    # via authlib
decorator==5.2.1
    # via ipython
dj-database-url==2.3.0
    # via -r requirements/requirements.txt
django==5.1.7
    # via
    #   -r requirements/requirements.txt
    #   dj-database-url
    #   django-ckeditor
    #   django-cors-headers
    #   django-extensions
    #   django-health-check
    #   django-js-asset
    #   django-solo
    #   django-storages
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   model-bakery
django-ckeditor==6.7.2
    # via -r requirements/requirements.txt
django-cleanup==9.0.0
    # via -r requirements/requirements.txt
django-cors-headers==4.7.0
    # via -r requirements/requirements.txt
django-extensions==3.2.3
    # via -r requirements/requirements.txt
django-health-check==3.18.3
    # via -r requirements/requirements.txt
django-js-asset==3.1.2
    # via
    #   -r requirements/requirements.txt
    #   django-ckeditor
django-solo==2.4.0
    # via -r requirements/requirements.txt
django-storages==1.14.5
    # via -r requirements/requirements.txt
django-stubs==5.1.3
    # via -r requirements/requirements.dev.in
django-stubs-ext==5.1.3
    # via django-stubs
djangorestframework==3.15.2
    # via
    #   -r requirements/requirements.txt
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.0
    # via -r requirements/requirements.txt
dparse==0.6.4
    # via
    #   safety
    #   safety-schemas
drf-spectacular==0.28.0
    # via -r requirements/requirements.txt
exceptiongroup==1.2.2
    # via
    #   -r requirements/requirements.txt
    #   ipython
    #   pytest
execnet==2.1.1
    # via
    #   -r requirements/requirements.txt
    #   pytest-xdist
executing==2.2.0
    # via stack-data
filelock==3.16.1
    # via safety
gunicorn==23.0.0
    # via -r requirements/requirements.txt
idna==3.10
    # via requests
inflection==0.5.1
    # via
    #   -r requirements/requirements.txt
    #   drf-spectacular
iniconfig==2.0.0
    # via
    #   -r requirements/requirements.txt
    #   pytest
ipython==8.34.0
    # via -r requirements/requirements.dev.in
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via safety
joblib==1.4.2
    # via nltk
jsonschema==4.23.0
    # via
    #   -r requirements/requirements.txt
    #   drf-spectacular
jsonschema-specifications==2024.10.1
    # via
    #   -r requirements/requirements.txt
    #   jsonschema
kombu==5.5.0
    # via
    #   -r requirements/requirements.txt
    #   celery
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
marshmallow==3.26.1
    # via safety
matplotlib-inline==0.1.7
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
model-bakery==1.20.4
    # via -r requirements/requirements.txt
mypy==1.15.0
    # via -r requirements/requirements.dev.in
mypy-extensions==1.0.0
    # via mypy
nltk==3.9.1
    # via safety
packaging==24.2
    # via
    #   -r requirements/requirements.txt
    #   dparse
    #   gunicorn
    #   marshmallow
    #   pytest
    #   safety
    #   safety-schemas
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.1.0
    # via -r requirements/requirements.txt
pluggy==1.5.0
    # via
    #   -r requirements/requirements.txt
    #   pytest
prompt-toolkit==3.0.50
    # via
    #   -r requirements/requirements.txt
    #   click-repl
    #   ipython
psutil==6.1.1
    # via safety
psycopg2-binary==2.9.10
    # via -r requirements/requirements.txt
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   safety
    #   safety-schemas
pydantic-core==2.23.4
    # via pydantic
pygments==2.19.1
    # via
    #   ipython
    #   rich
pyjwt==2.9.0
    # via
    #   -r requirements/requirements.txt
    #   djangorestframework-simplejwt
pytest==8.3.5
    # via
    #   -r requirements/requirements.txt
    #   pytest-cov
    #   pytest-django
    #   pytest-xdist
pytest-cov==6.0.0
    # via -r requirements/requirements.txt
pytest-django==4.10.0
    # via -r requirements/requirements.txt
pytest-xdist==3.6.1
    # via -r requirements/requirements.txt
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/requirements.txt
    #   celery
pyyaml==6.0.2
    # via
    #   -r requirements/requirements.txt
    #   drf-spectacular
referencing==0.36.2
    # via
    #   -r requirements/requirements.txt
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via nltk
requests==2.32.3
    # via safety
rich==13.9.4
    # via typer
rpds-py==0.23.1
    # via
    #   -r requirements/requirements.txt
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.10
    # via
    #   safety
    #   safety-schemas
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
ruff==0.10.0
    # via -r requirements/requirements.dev.in
safety==3.3.1
    # via -r requirements/requirements.dev.in
safety-schemas==0.0.11
    # via safety
setuptools==76.0.0
    # via safety
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   -r requirements/requirements.txt
    #   python-dateutil
sqlparse==0.5.3
    # via
    #   -r requirements/requirements.txt
    #   django
stack-data==0.6.3
    # via ipython
tomli==2.2.1
    # via
    #   -r requirements/requirements.txt
    #   coverage
    #   django-stubs
    #   dparse
    #   mypy
    #   pytest
tqdm==4.67.1
    # via nltk
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typer==0.15.2
    # via safety
types-pyyaml==6.0.12.20241230
    # via django-stubs
typing-extensions==4.12.2
    # via
    #   -r requirements/requirements.txt
    #   asgiref
    #   dj-database-url
    #   django-solo
    #   django-stubs
    #   django-stubs-ext
    #   ipython
    #   mypy
    #   pydantic
    #   pydantic-core
    #   referencing
    #   rich
    #   safety
    #   safety-schemas
    #   typer
tzdata==2025.1
    # via
    #   -r requirements/requirements.txt
    #   celery
    #   kombu
uritemplate==4.1.1
    # via
    #   -r requirements/requirements.txt
    #   drf-spectacular
urllib3==2.3.0
    # via requests
vine==5.1.0
    # via
    #   -r requirements/requirements.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via
    #   -r requirements/requirements.txt
    #   prompt-toolkit
